Added support for MacOS

+++ scripts/make_disk.sh
@@ -10,7 +10,12 @@ echo "Creating disk image"
 dd if=/dev/zero of=../build/disk.img bs=1M count=64
 
 echo "Formatting"
-mkfs.fat -F 32 ../build/disk.img
+
+if [[ "$(uname -s)" == "Darwin" ]]; then
+	mformat -i ../build/disk.img -F ::
+else
+	mkfs.fat -F 32 ../build/disk.img
+fi
 
 echo "Copying files"
 dd if=../build/stage1/boot.bin of=../build/disk.img bs=1 seek=0 count=3 conv=notrunc
